update: PUT coupons to have couponId in request params
